backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
PyMuPDF==1.24.0
google-generativeai==0.7.0
elevenlabs==1.5.0
pydub==0.25.1
aiofiles==24.1.0
email-validator>=2.0.0
groq>=0.9.0
gTTS>=2.5.0
